requirements.txt
python-binance
TA-Lib
numpy
websocket_client
pprint
config
json
requests
time
backtrader
matplotlib